# Remove on-the-fly memory allocation from WMP (20% active cycle speedup on depleted lattice)