Hey everyone,

I was recently doing a little dive on how WMP works, getting ready to put it on GPU, and found that there is allocation of a `std::vector` in the evaluation routine for WMP. Obviously, this is creating a pretty big performance hit. I found that since only up to ten polynomial terms are needed (which is only Gd157 IIRC), this is more appropriate as a `std::array`.

For a test problem, I used a 17x17 depleted lattice where every pin is at a different temperature. Obviously, it's pretty heavy on XS lookup. Here are my active tracking rates using 300,000 particles per generation, over one inactive and one active generation.


| Pointwise XS Interpolation | WMP (original) | WMP (new) |
|------------------------------------|----------------------|-----------------|
| 4372 | 3494 | 4153 |

I observed only a 6% speedup during inactive cycles, but am seeing 20% speedup in tracking in active. Nice. I imagine this is going to be helpful to event-based mode as well, but haven't tested it there.