@liangjg 2021-01-11
@gridley Nice improvement!
You're right, a maximum of 10th curve-fit order is used to generate the WMP library (check [the wmp library overview](https://github.com/mit-crpg/WMP_Library/blob/master/nuclides.md)).
But for the 10th order, there should be 11 coefficients so the `MAX_POLY_COEFFICIENTS` should be `11` rather than `10`?