@gridley 2021-01-11
Hey @liangjg, good catch. You're right about the WMP library going beyond 9th order. I had simply run with all the nuclides that had shown up from the depletion chain I used, and missed those 10th order ones as a result. I've now revised to handle anything up to 10th order.